BJ-134: Move util classes to common lib
