+++ pom.xml
@@ -18,6 +18,7 @@
     <modules>
         <module>serializer</module>
         <module>migrator</module>
+        <module>common</module>
     </modules>
 
     <dependencyManagement>
@@ -30,6 +31,11 @@
             <dependency>
                 <groupId>com.rbkmoney.geck</groupId>
                 <artifactId>serializer</artifactId>
+                <version>${project.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>com.rbkmoney.geck</groupId>
+                <artifactId>common</artifactId>
                 <version>${project.version}</version>
             </dependency>
             <dependency>
