+++ serializer/pom.xml
@@ -15,6 +15,10 @@
 
     <dependencies>
         <dependency>
+            <groupId>com.rbkmoney.geck</groupId>
+            <artifactId>common</artifactId>
+        </dependency>
+        <dependency>
             <groupId>com.rbkmoney.thrift</groupId>
             <artifactId>libthrift</artifactId>
         </dependency>
