+++ serializer/src/main/java/com/rbkmoney/geck/serializer/kit/msgpack/MsgPackProcessor.java
@@ -1,5 +1,6 @@
 package com.rbkmoney.geck.serializer.kit.msgpack;
 
+import com.rbkmoney.geck.common.util.StringUtil;
 import com.rbkmoney.geck.serializer.exception.BadFormatException;
 import com.rbkmoney.geck.serializer.StructHandler;
 import com.rbkmoney.geck.serializer.StructProcessor;
@@ -13,7 +14,6 @@ import org.msgpack.value.ValueType;
 import java.io.IOException;
 
 import static com.rbkmoney.geck.serializer.kit.EventFlags.*;
-import static com.rbkmoney.geck.serializer.kit.msgpack.StringUtil.expandAsciiString;
 
 /**
  * Created by vpankrashkin on 07.02.17.
@@ -176,7 +176,7 @@ public abstract class MsgPackProcessor<S> implements StructProcessor<S> {
         if (key > Character.MAX_VALUE) {
             throw new BadFormatException("Dictionary key is too long: "+ key);
         }
-        String str = expandAsciiString(data);
+        String str = StringUtil.expandAsciiString(data);
         if (dictionary.putIfAbsent((char) key, str) != null) {
             throw new BadFormatException("Dictionary key is already set: " + key);
         }
